Changed type of speed from NSInteger to CGFloat

+++ Circles/XYZAnimation.h
@@ -14,6 +14,6 @@
 
 - (NSInteger) animationID;
 - (NSInteger) minApplicableLevel;
-- (void) animate: (NSArray *) circles withSpeed: (NSInteger) speed;
+- (void) animate: (NSArray *) circles withSpeed: (CGFloat) speed;
 
 @end
